Code/GCP CF/centros_comerciales/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==1.4.2
google-cloud-storage==2.3.0
gcsfs==0.6.2
pymysql==1.0.2
sqlalchemy==1.4.0